first draft of essay
This is an initial commit of an essay draft on the topic of contributing to open source software.

It is organized as an rmarkdown document which will compile to github markdown, so that a compiled version of the file will be visible with the repository.  The rmarkdown document may also be compiled to pdf or other output modalities.